codebox-cli/pyproject.toml
[project]
name = "codebox-cli"
version = "0.1.0"
description = "CLI client for codebox sandbox containers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click",
    "docker",
    "websockets",
    "prompt-toolkit>=3.0.52",
    "rich>=14.3.3",
    "python-dotenv",
]

[project.scripts]
codebox = "codebox_cli.cli:cli"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/codebox_cli"]
